refactor: move termcode and integration to internal directory
Reorganize internal modules for better clarity by moving them to the
internal directory where they belong.

- Move lua/aibo/termcode.lua to lua/aibo/internal/termcode.lua
- Move lua/aibo/integration/init.lua to lua/aibo/internal/integration.lua
- Update all require statements to use new paths
- Keep individual integration files (claude, codex, ollama) in integration/
